Add extra_tests to xcode_workspace_config
Summary:
Support an optional `extra_tests` field in `xcode_workspace_config`. If
present, we pass down the extra test rules to the workspace and scheme
generator, which will add them and their dependencies.

Test Plan:
Created an xcode_workspace_config rule by hand with
multiple test targets extra_tests and no src_target. Ran
`buck project --workspace-and-projects --with-tests //Path/To:workspace`.

Confirmed generated workspace included TestActions for each of the tests,
and that the tests and their dependencies were marked as `buildForTesting` and
not `buildForRunning`.