Avoid deadlocks when running junit tests with custom timeouts.
Summary:
Occasionally we see deadlocks when running junit tests. It turns out that
in these cases, we're scheduling two tests that depend on each other to
complete on the same _single_ thread executor. As an example:

  suite -> testcase

The suite cannot finish until the testcase is complete, but testcase cannot
start until suite is completed. Fortunately, each test is created on a
separate thread to the main one, and so we can use a ThreadLocal to container
the executor. This means that child tests will each have their own single
thread executor, and if they spawn children, they'll have their own single
thread executor too.

Test Plan:
Add test that demonstrates deadlock
buck test --all