Genrule: include output name in rule key
Summary:
Previously, the output name used to form the output path used in
the genrule was not included in the genrule, which meant if the
user changed the output name in their BUCK file, buck wouldn't
necessary re-run the rule and generate the new output.

Test Plan:
Included a unittest.  Also, created a simple genrule locally and
verified that just changing the output name did not cause a rebuild
without this change.