Use ng-stop and ng-stats commands to kill buckd.
Summary:
Earlier, we used `os.kill()` to kill the buckd process by pid, and wait for 10
seconds to check if a process with the same pid is still running before
terminating the process using `SIGKILL`.

Now, we use the `ng-stop` command to shutdown the nailgun server which blocks
until the server is shutdown.

Test Plan:
```
$ buckd
$ buckd --kill
```
Verify buckd is killed.