Fail on errors to compile nailgun
Summary:
Ant defaults to `failonerror="false"` for the `exec` command, and we
did not set it to true.  This meant that we would happily ignore the
compilation failure and carry on, but this is bad if you happen to
already have buckd running and this is part of an update.

This change also makes us not try to build the client unless you are
on OS X or Linux, the two places we currently support nailgun anyway.

Test Plan:
* `ant client -verbose` works on both my mac and my dev server
* smoke builds