Enable 'buck project --profile' to profile buck.py performance
Summary:
We've gotten some complaints that buck.py performance is
unacceptable. This adds a new option to `buck project`, `--profile`,
which enables Python profiling.

If this option is specified, we turn on Python's cProfile profiler and
write to the debug logs the top 25 function calls by cumulative time.

When I enable this, I see that glob() is a top user of CPU time,
so tackling that first will give us lots of improvements.

Test Plan: `buck project --profile`, confirmed output in logs