Amend build to avoid invoking gcc on Windows.
Summary:
Most Windows installations don't have gcc installed, and those that do
tend to only run it in some UNIX emulation layer which only serves to
confuse and confound developers. Since nailgun isn't required to build,
skip the step that builds it if we're on Windows.

Test Plan: Run "ant" on a Windows machine.