Make ParamInfo stateless.
Summary:
Generating ParamInfo is an exercise in reflective goodness/madness in Java.
For each and every BuildRule in the Action Graph, we walk every field of
the ConstructorArg and create a ParamInfo, which is then called to coerce
the value from the JSON map to the arg. Needless to say, this is grossly
inefficient, and is the main reason why we keep the raw Map of values around
in the BuildRuleFactoryParams (that is, we sacrifice space for speed, since
there's no way to tell whether or not we'll need the coerced TargetNode
values or not).

We can speed up parsing by reducing the amount of reflection done, and this
seems like low-hanging fruit. Rather than doing all the changes in one fell
swoop, I'm going to break things up. This change simply allows ParamInfo
instances to be reused with any instance of a ConstructorArg, rather than
just the ones from a particular build file.

Test Plan: buck test --all