Get the Buck scripts running under Windows.
Summary:
In particular, this now means that it's possible to invoke:

  python bin\buck --help

and get the buck help text. Since there's a chance that git may not be on the
PATH we check that it's present before trying to figure out whether the binary
is present before attempting to get the git hash.

Test Plan: python bin\buck --help