Remove the webkit-specific webkitCreateShadowRoot() with createShadowRoot().
Summary:
Our fork of the Chrome Trace is a bit dated, but last I checked, it is still
buggy: https://code.google.com/p/trace-viewer/issues/detail?id=478. As you
can see, the issue is marked closed, but my repro case demonstrates that it
is still buggy. Therefore, instead of updating to a new version to fix this
bug, I just modified the code and it seems to work fine.

Test Plan: Verify that a trace under the `/trace` URI in the webserver loads again.