Combine generating R.java and packaging resources in the same aapt invocation.
Summary:
Previously, we used `UberRDotJava` to generate the uber R.java file using `aapt`
and `AaptPackageResources` to generate the resource APK, i.e. APK containing
the compiled resources and assets. This was done so that we could compile and dex
the R.java as soon as it was available and not block dependent rules. That is no
longer a restriction with pre-dexing.

This diff merges the two into `AaptPackageResources`. As a side-effect, this
generally reduces build times because we only have one aapt invocation running
(even though previously the 2 invocations were run in parallel).

Test Plan: buck test test/com/facebook/buck/android