Create a web page to see the output from StreamingWebSocketServlet.
Summary:
`WebServerBuckEventListener` and `StreamingWebSocketServlet` work together to push
Buck events down to a client via a WebSocket. This could be used for:

* Creating an HTML5 SuperConsole.
* Creating a realtime visualization of a build using d3: https://github.com/mbostock/d3/wiki/Gallery.
* Creating a communication channel with an IntelliJ plugin.

As such, it is useful to have a lightweight way to test out the WebSocket.

Test Plan:
Went to http://localhost:XXXXX/static/test_websocket.html while a build was
running, opened the JavaScript console, and verified that events were printed to
the console, as expected.