When port=0 in .buckconfig for [httpserver], it will pick a random free port.
Summary:
If multiple checkouts of the same repository are being built on the same machine
with Buck simultaneously, then it is important that the corresponding webservers
do not try to claim the same port. In that case, it makes sense to configure
`.buckconfig` as follows:

```
[httpserver]
  port = 0
```

If individuals want to use a stable port, then they can override this in `.buckconfig.local`.

They can also use a negative value for `port` to disable the web server.

As part of this diff, this is now documented in `docs/concept/buckconfig.soy`.

Test Plan:
Used `port = 0` and verified that a random port was chosen and was printed out to the
console by Buck. When `buckd --kill` was run, verified that the webserver was shut down.