Ensure we print buck.py stderr to logs
Summary:
Currently, `ProjectBuildFileParser` ensures it writes
buck.py's stderr to `Console`, but nothing ensures stderr goes to the
logs.

I've been working on improving `buck.py`'s `glob()` method, and it was
really difficult to diagnose problems via Buck's debug logs.

To help with that, this diff introduces an optional delegate
`InputStreamConsumer.Handler` which allows users of
`InputStreamConsumer` to get a callback for each line of input.

I use the delegate from `ProjectBuildFileParser` to log a warning
to the debug logs for each line printed by buck.py to stderr.

(Unlike error logs, we do not send warning logs to the console -- only
to the debug logs on disk.)

Test Plan:
Added `print >>sys.stderr, "Oops"` to `buck.py`. Confirmed
it showed up on console and in the logs.