Allow android binary rules to depend on C/C++ libraries
Summary:
This diff connects android rules up to our C/C++ rules, by making the
from-source and prebuilt C/C++ rules implement the ##AndroidPackageable##
interface.  The supported NDK C/C++ platforms are threaded through
the ##AndroidBinaryDescription## and are used when constructing the
collector, which C/C++ libraries use to find the C/C++ platforms they
need to build the shared libs against.

Test Plan:
Added a few tests for the prebuilt and from-source C/C++ interfaces
and an integration tests for an android binary depending on a C/C++
library.