Parser: unify recursive and explicit target graph parsing
Summary:
Currently, we support two ways of parsing and constructing the
target graph:
1) Parse starting from a list of explicit targets.
2) Recursively search for all build targets in the repo.
The former was mostly done in java, with the buck.py python parser
handling parsing individual build files.  The latter delegated to
the buck.py python parser to perform the recursive search of all
build files and then parse them and return them as one blob.

This diff unifies the target graph parsing around a single method
in the parser, which uses new target node specification objects to
describe the various parsing styles we support, which is currently
just a recursive search based on ##RuleJsonPredicate## and an exact
match base on explicitly specified ##BuildTarget## objects.  In both
cases, the major lifting is all done in the parser, including
recursive parsing, which showed to be a lot faster than it's python
equivalent (this made "buck targets" twice as fast in fbcode).

This also opens the door to introduce new specifications to control
how the target graph is built.  In particular, it's be interesting
to support some UI to build or list targets for all packaging rules
under a particular directory (which would be really useful for
automation).

Test Plan: buck test --all