Add test for BuildTargetTypeCoercer when dealing with Windows paths.
Summary:
This has broken twice. Time to add a test. Of course, it's not a simple thing
since EasyMock won't mock out `toString`. As such, we use a dynamic proxy from
normal Java. The other problem was that the test wouldn't ever pass on a
UNIX-like OS since the path separator is a "/", so the path was being returned
unmodified. We now brute force the separator to always be "/", though it's not
terribly elegant.

Test Plan: buck test --all