Remove SourcePath#getName.
Summary:
To remove BuildRule creation as a prerequisite of TargetGraph
population, the SourcePath implementation referring to rules
must be changed from taking a BuildRule to taking a BuildTarget.
This moves the logic to determine the name of an output file
referred to by a SourcePath from the SourcePath to helper
methods to prepare for that.

Test Plan: CI