Add a `FlavorDomain` class to organize flavors
Summary:
It's common to use flavors, which are essentially boolean settings,
to model types with larger domains.  This diff adds ##FlavorDomain##
class to make this easier.  ##FlavorDomain## objects are constructed
around a mapping that defines their entire domain in terms of flavors
and maps them to some user-defined value.  An example of a domain
would be something like the various thrift languages that support,
mapped to the enhancers the thrift description uses to construct
language specific rules.  As part of this diff, thrift was also
converted to use this domains.

This diff also provides a helper function to propagate flavors,
represented by ##FlavorDomain## objects from a dependent target to
dependent targets.

Test Plan: included tests