Make AbstractAppleNativeTargetBuildRule responsible for its header map location.
Summary:
The location of a header map seems like something that should be part of the
rule rather than known only to `ProjectGenerator`. This will be even more important
when we teach Buck to build iOS.

The various header map suffixes that were hardcoded as strings in `ProjectGenerator`
now appear as an enum in `AbstractAppleNativeTargetBuildRule`.

In order to do this refactoring, I had to pass a `BuildRule` as a stronger type,
`AbstractAppleNativeTargetBuildRule`, as appropriate, in `ProjectGenerator`.
Previously, much of this logic was based on a `BuildTarget` that could have come
from any `BuildRule`.

Also moved header maps from `buck-out/path/to/target` to `buck-out/gen/path/to/target`,
as that seemed more appropriate. We do not want to have arbitrary subdirectories
in `buck-out`.

Test Plan:
ProjectGeneratorTest
Sandcastle