Move macros expander logic to it's own class for reuse
Summary:
This diff strips out the macro expansion logic from the
`AbstractGenruleStep` step and moves it into it's own library so
that it's usable by other rules and more extensible.  The main class
is now `MacroHandler`, which supports two operations to identify
build target references from macros a string (for use in the
`findDepsFromParams` method) and to expand macros in a string.
In both cases, the `MacroHandler` is responsible for parsing out
macros of the form $(<macro>[ <optional-arg>]).  The `MacroHandler`
is constructed with a map of macro names to `MacroExpander` classes
which are responsible for defining the operation of particular
macros (e.g. a there is a "location" macro expander).  The `MacroExpader`
classes are passed the optional arg extract by the `MacroHandler` and
support operations to both expand it and to resolve and `BuildTarget`
references.

Test Plan:
Added some tests and restructured some existing ones for the new
setup.