Preprocess and compile C/C++ sources in separate rules
Summary:
Previously, we had been preprocessing and compiling C/C++ sources
in the same build rule.  This diff separates them out into separate
steps and build rules, which allows the same preprocess source to
share the cache and be reused for both PIC and non-PIC compilations,
and to also be reused for any static analysis tools.

Test Plan: Added tests.