[mini-aapt] Do not look for xml id definitions/usages in xml files under values* directories.
Summary:
I introduced test failures in my previous attempt. This diff fixes the failures,
and also uses the same condition to check if a directory is a valid values dir
in both the places.

For a sample resource module with 100 strings per locale and about 40 locales,
this halves mini-aapt's execution time.

~670s vs ~300s

Test Plan: buck test --all