Add methods to retrieve rule and raw path to SourcePathResolver.
Summary:
Getting a rule from a SourcePath that refers to the output of one
is trivial right now, but it will require a resolver when we
change the SourcePath to use a BuildTarget instead of a BuildRule.
This adds methods to SourcePathResolver to retrieve the rule whose
output a SourcePath refers to, and the raw path in the case it
doesn't refer to any rule's output.

Test Plan: CI