Remove JavaLibrary#isVisibleTo.
Summary:
JavaLibrary#isVisibleTo is only used to checking suggested dependencies,
but introduces the visibility information, which shouldn't be needed in
either the target graph or the action graph, into the action graph.
While checking the target's visibility is a good thing it's not critical
enough to warrant this extra complication.

Test Plan: CI