Improve caching logic for AbstractAppleNativeTargetBuildRule.
Summary:
Implemented `getInputsToCompareToOutput()` and `appendDetailsToRuleKey()`
for `AbstractAppleNativeTargetBuildRule`. Because it has a field that is
a `GroupedSource`, there is now a `GroupedSource.Visitor` that can be
used to accumulate all of the data from the nodes in a `GroupedSource`
tree in a consistent order (assuming the order in which groups and sources
are passed from a BUCK file into Buck is consistent, which might not be
true if `glob()` is involved).

Test Plan: Sandcastle builds.