Add a mechanism for generating stub jars.
Summary:
This puts in place a piece of infrastructure for generating ABI-stable jars
from class files. These jar files contain just enough bytecode to allow java
to be compiled against them, but it's extremely inadvisable to attempt to use
them.

The stub-jar generated has the following features per class:

  * Fields are ordered in alphabetical order.
  * Methods are ordered in alphabetical order.
  * Private methods and fields are ignored.
  * No method has a method body.
  * Bridge and synthetic methods inserted by the compiler are ignored.

In addition, the jar is generated without timestamps, so it should be possible
to run this over functionally equivalent inputs and generate the exact same
output.

Test Plan: buck test --all