Add an optional prefix_header argument to apple_library() and apple_binary().
Summary:
This diff does not make use of the `prefix_header` argument yet: it just makes it
available via `AbstractAppleNativeTargetBuildRule`. Ultimately, it should be used
for building.

In a `.pbxproj`, it is possible to specify a `.pch` file for use with the `-include`
param via `GCC_PREFIX_HEADER = <path-to-file>`. Internally, we often do this via the
macro in `Configurations/Xcode/Projects/Project.xcconfig`:

    GCC_PREFIX_HEADER = $(TARGET_NAME)/$(TARGET_NAME)-Prefix.pch

My migration plan is to take our `ios_library()` and `ios_binary()` macros
and do something like:

```
def ios_library(
    name,
    pch_file = 'PLACEHOLDER_VALUE',
    ...
    ):
  if pch_file == 'PLACEHOLDER_VALUE':
    pch_files = glob(['**/*.pch'])
    pch_file = find_pch_file_closest_to_root_if_one_exists(pch_files)

  apple_library(
    name = name,
    prefix_header = pch_file,
    ...
```

This way, if a `pch_file` is specified for `ios_library()`, it will be used
as the `prefix_header` for the `apple_library()`.
Similarly, if `pch_file` is explicitly specified to be `None`, then the
`prefix_header` for the `apple_library()` will be `None`.
If `pch_file` is unspecified, then we use a heuristic to try to find the
appropriate value for `pch_file` in the interim while we go through
the painful process of updating all of our existing `ios_library()` rules.
Once we have updated all of our rules, we can remove the heuristic.

Test Plan: Sandcastle builds.