Add a `NdkCxxPlatform`
Summary:
This adds a `CxxPlatform` to encapsulate and android NDK and uses
it to build C/C++ rules.  It's parameterized by the NDK root path,
the SDK ABI (e.g. armeabi, x86), and the host platform, via it which
it finds all the correct locations in the NDK.

Currently, I've added implementations for armeabi, armeabi-v7, and
x86, and added them to the available C/C++ flavors, so that we can
build C/C++ binaries and tests against an Android NDK.  This isn't
by itself that useful, but it does allow some form of testing that
builds work.

Test Plan:
Added an integration in C/C++ binary that builds against the
available NDK platforms.