Lazily evaluate `xcode-select --print-path` and make it available via the ExecutionContext.
Summary:
Instead of assuming the developer is using `/Applications/Xcode.app/Contents/Developer` as
his or her Apple developer directory, find the location using `xcode-select --print-path`.

Test Plan:
Exercised this logic locally on a Mac with `xcode-select` installed.
Sandcastle builds.

It is difficult to test `ExecutionContext#createAppleDeveloperDirectorySupplier()` directly
because either:
(1) It has to be run on a Mac where the location of the developer directory is known.
(2) So much of the logic would have to be mocked out that there would be nothing meaningful left to test.