Introduce immutable value type autogeneration
Summary:
Currently, we have to manually write immutable value types,
including implementing `hashCode()`, `equals()`, `description()`, and
(if we want to avoid an explosion of constructor params), a `Builder`.

This is cumbersome, time-consuming, and causes bugs (see D1495875 for
an example).

This diff introduces the Immutables annotation processor and code generation
tool from http://immutables.github.io/ under `//third-party/java/immutables`.

I updated both the ant and Buck builds to make use of it and included
a test under `//test/com/facebook/buck/immutables`.

Note that it's currently a bit annoying to write a BUCK file to use
the processor (you have to explicitly specify the annotation processor
class, the dep for the annotation processor JAR, and the deps of the
generated classes).

We can fix this with a utility function (`java_immutables_library`?)
which automatically adds the parameters, but I'm open to other suggestions.

Also, the ant build currently has to run:

 javac -processorpath immutables/generate-tool.jar -proc:only

as a separate step to generate the source files. It does so separately
for source and test directories, but this does add a few seconds to
the ant build. (I tried doing this with the Eclipse compiler, but
it seems to ignore the `-processorpath` parameter entirely.)

I fixed the IntelliJ IDE files to handle this new dependency, but not the Eclipse files. @mbolin said he'd take care of Eclipse.

Test Plan:
  ant test -Dtest.class=TypeTest
  buck test //test/com/facebook/buck/immutables:test

Ran new test via IntelliJ IDE UI.