Unify thrift:compiler_path and thrift:compiler_target
Summary:
Previously, we used two different buck config options depending on
whether thrift was generated by a build rule or was an existing file.
This unifies to use a single ##thrift:compiler## field and parses
based on whether it looks like a build target or not.

Test Plan: updated tests; ran them