refactor ProjectGenerator path relativization to helper class
Summary:
The ProjectGenerator often relativizes paths that are based on the buck project
root (repo root) to one based on the output directory (where the xcodeproj is
generated).

This diff refactors logic related to that into a separate helper class.
Hopefully what the relativization functions are clearer than before.

This will also be useful for a future diff where I break up some of the
functionality of ProjectGenerator into a different class.

Test Plan: generate and build wilde, existing tests