Don't parse target graph twice in `PartialGraph.createFullGraph`
Summary:
Previously, this method would first create a target graph containing
every node to find the `BuildTarget` names for everything in the repo.
Then, it's use this list of `BuildTarget` names to generate a huge
lit of (all-inclusive) filters to use in the second pass to build a
target graph.  This second pass was especially costly, as it spent
a *lot* of time using the filters to extract the build targets to
match from the build files.

This diff simplifies the process: the all-inclusive target graph is
parsed once, and the full list of target node names are pulled directly
from that.

Test Plan: `buck targets`