ProjectsCommand: build target graphs less frequently
Summary:
The projects command would construct a target graph twice: once to find the
name of all nodes, and a second time to actually get a target graph.
The second pass was extremely expensive since it had a lot of filters
to process.  This diff reduces this to a single pass.

Test Plan: ##buck project##