Initial work on generating compilation database for libclang.
Summary:
This adds a `#compilation-database` flavor to `apple_library()` and `apple_binary()`.
If you run:

```
buck project --ide xcode
buck build //apple/library:rule#compilation-database
cat `buck targets --show_output //apple/library:rule#compilation-database`
```

Then the clang calls in the database should succeed if the project is simple enough.
The clang flags are a subset of what is ultimately necessary. Further, they are currently
hardcoded to use settings for iOS 8.

Unfortunately, this diff required adding a lot of similar code to `AppleLibraryDescription`
and `AppleBinaryDescription`. I believe that the roles of these rules are too similar right
now: they should have a clearer distinction so that we don't have to support all of the
same flavors for both rules.

Test Plan:
Invoke manually on small projects that use header maps and
verify that the generated command exits with a non-zero exit code.

Created `CompilationDatabaseIntegrationTest`. Ran:

  buck test --code-coverage --code-coverage-format html --no-results-cache test/com/facebook/buck/apple/

Opened the following files in the browser to inspect the code coverage:

  buck-out/gen/jacoco/code-coverage/com.facebook.buck.apple/AbstractAppleNativeTargetBuildRuleDescriptions.java.html
  buck-out/gen/jacoco/code-coverage/com.facebook.buck.apple/AppleBinaryDescription.java.html
  buck-out/gen/jacoco/code-coverage/com.facebook.buck.apple/AppleLibraryDescription.java.html
  buck-out/gen/jacoco/code-coverage/com.facebook.buck.apple/CompilationDatabase.java.html

It looks quite strong!