Add support for `buck build --keep-going` a la Make.
Summary:
When `--keep-going` is specified, a parse-friendly report is printed to stderr,
detailing which targets were built successfully.
Each line of the report represents the status of one build target, which has up to four columns:
* `OK` or `FAIL`, as per the success of the build rule.
* build target of the rule.
* If successful, the type of the success from `com.facebook.buck.rules.BuildRuleSuccess.Type`.
* If successful, the path to the output file of the rule, if it exists.

Here is a possible sample output for `buck build --keep-going //fake:rule1 //fake:rule2 //fake:rule3`:
```
OK   //fake:rule1 BUILT_LOCALLY buck-out/gen/fake/rule1.txt
FAIL //fake:rule2
OK   //fake:rule3 FETCHED_FROM_CACHE
```

When `--keep-going` is specified, the exit code will be 0 only if all targets were
built successfully.

The familiar `-k` alias for `--keep-going` is not supported because
`-k` is currently used as an option for `buck uninstall` as `adb uninstall`
supports a `-k` flag to keep the data and cache directories during uninstallation.
Note that `buck install` and `buck uninstall` inherit all of the flags
that `buck build` supports, which is why there is a collision.
If we decide this inconsistency with Make is problematic, then we can
consider renaming the `-k` flag to `buck uninstall` at a future date.

Test Plan: Added `BuildCommandTest` and `BuildKeepGoingIntegrationTest`.