Add utility to BuckConfig for resolving paths that may be outside the project root.
Summary:
This is exposed via `BuckConfig` so that it will be available in `AbstractCommandOptions`.
It appears that `org.kohsuke.args4j.spi.PathOptionHandler` does not resolve relative path inputs,
so commands with options that take paths that could be outside the project root need to
be able to resolve paths in a consistent way.

This method accepts null so that it is easier to use with optional `Path` params.

This will be used in an upcoming diff that adds a `--build-report` option to `buck build`.

Test Plan: Updated `BuckConfigTest`.