Modify --build-report to produce a report in an extensible JSON format.
Summary:
The format of report written by `--build-report` is:

```
{
  "results": {
    "//fake:rule1": {
      "success": true,
      "type": "BUILT_LOCALLY",
      "output": "buck-out/gen/fake/rule1.txt"
    },
    "//fake:rule2": {
      "success": false
    },
    "//fake:rule3": {
      "success": true,
      "type": "FETCHED_FROM_CACHE"
    }
  }
}
```

This is designed to be extensible, which is why the results are one layer
deep in the map. Also, each result is a map rather than a string so more
details about the build result can be added in the future without changing
the current schema.

Future extensions to this format could include build start/stop time
or information about why a rule failed to build.

Test Plan: Updated `BuildCommandTest` and `BuildKeepGoingIntegrationTest`.