Make buck project exclude subdirectories correctly.
Summary:
When generating intellij config for a module, buck is supposed to exclude all
subdirectories which contain their own BUCK files. This was broken when we
introduced InMemoryBuildFileTree (for perf); this fixes it by reverting back
to the build file tree that crawls the file system. Also verified that this
doesn't regress the project command execution time (because I assume we only
call buildFileTree.getChildPaths() for project_config rules).

Test Plan: Integration test.