Change XcodeRuleConfigurationTypeCoercer to accept a map.
Summary:
XcodeRuleConfigurationTypeCoercer coerces lists of maps, with each
map being coerced to an XcodeRuleConfigurationLayer. This requires
wrapping a map in a list even if a configuration has only one layer.
This changes XcodeRuleConfigurationTypeCoercer to treat a map as a
configuration with only one layer.

Test Plan: CI, new unit tests included