Add an escaper for makefile strings
Summary:
This adds an escaper for values placed on the right-hand side
of assignments in GNU-style makefiles, and is useful for makefile
generation for use with the Android NDK.

GNU makefiles use an interesting escaping strategy, where the
escape sequence (i.e. backslash) must itself be escaped *only*
if it precedes a character that needs to be escaped (e.g. so
a backslash in the original string won't necessarily be escaped).

Test Plan: added some tests; generated some Android.mk files