Update the exopackage support library to initialize ExopackageSoLoader.
Summary:
In order to support projects using the old "exopackage" boolean parameter of
the android_binary rule, keep the old constructors.

Load secondary dex files/initialize SoLoader only if the corresponding feature
is enabled in the build, which is indicated by the exopackage flags.

Test Plan:
Added a new activity to the test app that loads two shared libraries and calls
jni methods implemented in those.

jni/two:two is defined for x86 and armeabi, to test devices that support both
armv7 and arm.