Bump the Xcode upgrade checks version.
Summary:
Xcode upgrade checks are useful for normal Xcode projects because
a lot of the settings that are set in the project, even those by
Xcode itself, can become obsolete. In the case of Buck generated
projects, because the only settings that are set are the ones the
user or Buck has a need to specify, these checks are less useful.
This bumps the upgrade check versions to disable the warnings that
suggest updating the project settings.

Test Plan: CI, open a project with Xcode 6.1, observe no upgrade warnings appear