Skip handling of frames and code when generating stub jars.
Summary:
We have relied on ASM being able to parse _everything_ in a class file when generating
an abi jar. This is slower than it could be (since more data needs to be processed)
and causes a problem when ASM can't parse the code. Stop doing that, and go faster.

Test Plan: buck test --all