Write out build settings to a generated xcconfig file when possible.
Summary: When external xcconfigs are not referenced, we can optimize performance by writing Buck settings into a generated xcconfig and referencing that instead of embedding the settings in the project.

Test Plan: Generate projects that don't reference external xcconfigs. Confirm the generated xcconfigs look correct and are referenced correctly in the project.