Pass BuckEventBus to TargetGraph constructor instead of getActionGraph()
Summary:
I want to use `BuckEventBus` in the memoized `Supplier` in
`TargetGraph`, but that object is currently only passed to
`getActionGraph()`.

This moves the `BuckEventBus` out of `getActionGraph()` and into the
constructor of `TargetGraph` so the next diff can use it in the
`Supplier` instead of in `getActionGraph()`.

Test Plan: ant java-test -Dtest.class=ParserTest