C/C++: support sanitizing additional platform-specific paths
Summary:
This diff extends the `DebugPathSanitizer` so that it can sanitize
additional platform-specific, non-compilation-directory paths, and
adds an additional method to the `CxxPlatform` interface to return
the `DebugPathSanitizer` to be used for that platform.  This is
useful for platforms which rely on user-specified directories, such
as the `NdkCxxPlatform` which uses a NDK root directory given by the
user.

This diff also adds default implementations for the `DebugPathSanitizer`
instances returned from the platform objects, forcing the current
working directory to `.` and, in the case of `NdkCxxPlatform`, also
sanitizes the NDK root to `.`.  This prevents user-specific paths
from affect cachability, but will require a line in the user's .gdbinit
to locate the correct NDK root location (or user's can run a GDB command
when they start GDB).

Test Plan:
Included some tests.  Also built C/C++ code for the default and NDK
platforms and verified that `.` was used instead of user-specific
paths.