Tweak ExopackageInstaller to create symlinks to native library files.
Summary:
Currently, we copy native shared objects to the exopackage directory with file
names of the format "native-<sha1-hash>.so". This works fine as long as the only
code that loads libraries is in Java. However, let's say there are two libraries
A and B, where B depends on A, and assume that the Java code ensures that it
loads A before loading B; and that native code inside B also tries to load A.
This ends up being no-op because Java has already loaded A. However, this only
works if the file name that Java loaded A from is named "libA.so".

This diff addresses the problem by creating symlinks of the form "lib<name>.so"
to the actual files already copied in the exopackage directory. This also makes
the ExopackageSoLoader a lot simpler, because it doesn't need to read the
metadata txt file to build up a map from library name to the name containing
the sha1 hash.

Test Plan: ./script/exopackage_test_single.sh against a Galaxy Nexus.