New CxxPlatform implementation AppleCxxPlatform
Summary:
This is a new implementation of `CxxPlatform` to build iOS
and Mac OS X binaries. We'll use this to calculate the paths to
various binaries included with the Xcode SDK, as well as the
flags to pass to the compiler, assembler, linker, etc. to target
the appropriate architecture, SDK version, and so on.

It's mainly a skeleton for now; I haven't added the compiler / linker
flags yet.

I copied a fake layout of tools from Xcode 6.0 (6A313) with iPhone OS 8.0's
SDK. This version of the SDK doesn't include any tools in the SDK-specific
directory (`Platforms/iPhoneOS.platform/Developer/SDKs/iPhoneOS8.0.sdk/usr/bin`)
that we're interested in, but git doesn't allow empty directories, so I included
a fake `atos` so the directory isn't empty.

Test Plan: Unit test included.