Add a utility class for in-place byte array replacements
Summary:
This diff adds a `ByteBufferReplacer` classes that performs in-place
same-length find-and-replace operations on `ByteBuffer` objects.  The
main added value here is coercing the `Pattern` and `Matcher` classes
to work with arbitrary byte arrays, by converting the underlying bytes
to and from ISO-8859-1 code points, which map 1-to-1 from chars to bytes.
Since the `Matcher` class uses boyer-moore under-the-covers for `LITERAL`
strings, the find-and-replace can be done very quickly.

This class also provides a static helper method to operate on a map
of `Path` objects representing paths to be replaced in the `ByteBuffer`,
using the given `CharSet` to convert to the low-level byte arrays needed
by the class, filling with the provided path separator char when needed.

Test Plan: Added a test case for available character sets.