Exclude per file flags `public` and `private` from clang flags.
Summary:
Previously, we blindly included all per-file-flags to the list of clang flags.
Now, we have a blacklist that filters out `public` and `private`.

Test Plan:
Updated `CompilationDatabaseTest`.
Verified that a clang invocation that previously failed because it included `public`
in its list of flags now succeeds when it is removed.