Do not use a parser to filter nodes in the graph.
Summary:
During project generation we always read in the whole target graph so it is
unnecessary to use parser to filter nodes from the graph because we can just
filter them directly from the loaded graph. This diff removes an unnecessary
use of the parser.

Test Plan: `buck test --all`