Make Xcode prompt message include absolute path to repository
Summary:
Users are getting confused by the prompt in `ProjectCommand`
which says to edit `.buckconfig.local` but not that it needs to be
in the repo root.

This updates the message to use the repo root as a relative path,
and stops hard-coding `.buckconfig.local` as a string, instead
using it from `BuckConfig`.

Test Plan:
`buck project` in Apple repo while Xcode is running:

  % buck project
  Xcode is currently running. Buck will modify files Xcode currently has open, which can cause it to become unstable.

  Kill Xcode and continue? [Y/n]
  To disable this prompt in the future, add the following to /Users/bgertzfield/src/fbobjc/.buckconfig.local:

  [project]
    ide_prompt = false