Change BuildFileTree#getBasePathOfAncestorTarget to return an optional.
Summary:
BuildFileTree#getBasePathOfAncestorTarget cannot distinguish between a root
basePath, meaning a build file in the project root, and a missing base path,
meaning no build files in any of the ancestors of the path. This changes the
return type to be an optional so this distinction can be communicated to the
caller.

Test Plan: CI, new tests