Pass the java compiler environment around properly.
Summary:
This means that if you specify an external javac to use, it will now be
used to compile all java code, rather than some of it. This paves the way
for a follow up diff that allows users to set the bootclasspath for code
in their buckconfig.

Test Plan:
buck test --all
build the major facebook android apps successfully.